vera/backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
qdrant-client
google-generativeai
unstructured
pdfplumber
pydantic
